Use single table imports instead of schema object
By importing standalone table objects to use inside `.select().from()`
queries, we allow tree shaking to work for the actual tables used in
queries, since using whole schemas as in `schema.table` rather than
`table` inside `.from()` will result in inclusion of the whole schema
object, and schemas can get quite large as the app grows.

It is worth mentioning though, that this optimization is defeated if we
are using the query builder as in `db.query.table` because then, we will
have to pass the whole schema to the drizzle function during
initialization anyway (as in `drizzle(client, { schema })`).